+++ SkillModifiers/after_image.lua
@@ -1,7 +1,9 @@
-SkillModifier.register_modifier("after_image", 250, function(skill)
-    return SkillModifier.get_modifier_num("after_image") < 2
-end, function(skill, data)
-    SkillModifier.add_on_activate_callback(data, function(skill_)
-        gm.apply_buff(skill_.parent, 44.0, skill_.cooldown_base)
+local after_image = SkillModifierManager.register_modifier("after_image", 250)
+after_image:set_add_func(function (data, modifier_index)
+    data:add_pre_activate_callback(function (data)
+        gm.apply_buff(data.skill.parent, 44.0, data.skill.cooldown_base)
     end)
+end)
+after_image:set_check_func(function (skill)
+    return SkillModifierManager.count_modifier(skill, "after_image") < 2
 end)